Moved files out of the Infrastructure, they didn't belong there. Now infrastructure is DB implemention stuff.
