+++ app/EventStore/EventStore.php
@@ -1,13 +1,25 @@
 <?php namespace App\EventStore;
 
-interface EventStore
+class EventStore
 {
-    public function store(array $events);
+    private $event_repository;
+    private $event_stream_factory;
     
-    /**
-     * @param type $aggregate_type_id
-     * @param type $aggregate_id
-     * @return EventStream
-     */
-    public function fetch($aggregate_type_id, $aggregate_id);
+    public function __construct(
+        EventRepository $event_repository,
+        EventStreamFactory $event_stream_factory)
+    {
+        $this->event_repository = $event_repository;
+        $this->event_stream_factory = $event_stream_factory;
+    }
+    
+    public function store(array $events)
+    {
+        $this->event_repository->store($events);
+    }
+    
+    public function fetch(AggregateID $aggregate_id)
+    {
+        return $this->event_stream_factory->aggregate_id($aggregate_id);
+    }
 }
