+++ infrastructure/app/EventStore/PDO/EventRepository.php
@@ -3,9 +3,9 @@
 use PDO;
 use App\EventStore\Event;
 use App\EventStore\EventBuilder;
-use Infrastructure\App\EventStore\AggregateID;
+use App\EventStore\AggregateID;
 
-class EventRepository implements \Infrastructure\App\EventStore\EventRepository
+class EventRepository implements \App\EventStore\EventRepository
 {
     private $pdo;
     private $select_statement;
